remove Xwayland view
