+++ src/view.cpp
@@ -9,7 +9,7 @@
 const int DEFAULT_MINIMUM_WIDTH = 800;
 const int DEFAULT_MINIMUM_HEIGHT = 600;
 
-View::View(Controller *server_)
+View::View(Controller *server_, wlr_xdg_surface *surface)
   : mapped(false)
   , x(0)
   , y(0)
@@ -18,7 +18,9 @@ View::View(Controller *server_)
   , old_height(DEFAULT_MINIMUM_HEIGHT)
   , old_x(0)
   , old_y(0)
-  , server(server_) { }
+  , server(server_)
+  , xdg_surface(surface)
+{ }
 
 void View::focus_view(wlr_surface *surface) {
   wlr_seat *seat = server->seat;
@@ -53,29 +55,6 @@ bool View::tiled() const {
 bool View::maximized() const {
   bool maximised = state == WM_WINDOW_STATE_MAXIMIZED;
   return maximised;
-}
-
-void View::window(bool restore_position) {
-  bool is_windowed = windowed();
-  if (is_windowed) {
-    return;
-  }
-
-  if (restore_position) {
-    x = old_x;
-    y = old_y;
-  } else {
-    wlr_box geometry;
-    extents(&geometry);
-
-    float surface_x = server->cursor->x - x;
-    float x_percentage = surface_x / geometry.width;
-    float desired_x = old_width * x_percentage;
-    x = server->cursor->x - desired_x;
-  }
-
-  server->damage_outputs();
-  state = WM_WINDOW_STATE_WINDOW;
 }
 
 void View::toggle_maximized() {
@@ -206,4 +185,156 @@ void View::unmap_view() {
   mapped = false;
   server->focus_top();
   server->damage_outputs();
+}
+
+void View::committed() {
+  server->damage_output(this);
+}
+
+void View::for_each_surface(wlr_surface_iterator_func_t iterator, void *data) const {
+  if (xdg_surface->surface == NULL) {
+    return;
+  }
+  wlr_xdg_surface_for_each_surface(xdg_surface, iterator, data);
+}
+
+void View::focus() {
+  focus_view(xdg_surface->surface);
+}
+
+void View::unfocus() {
+  wlr_xdg_toplevel_set_activated(xdg_surface, false);
+}
+
+void View::geometry(struct wlr_box *box) const {
+  wlr_surface_get_extends(xdg_surface->surface, box);
+
+  if (!xdg_surface->geometry.width) {
+    return;
+  }
+
+  wlr_box_intersection(&xdg_surface->geometry, box, box);
+}
+
+const wlr_surface* View::surface() const {
+  return xdg_surface->surface;
+}
+
+void View::save_geometry() {
+  if (state == WM_WINDOW_STATE_WINDOW) {
+    wlr_box geometry;
+    extents(&geometry);
+
+    if (geometry.width != 0) {
+      old_width = geometry.width;
+      old_height = geometry.height;
+    }
+
+    old_x = x;
+    old_y = y;
+  }
+}
+
+void View::tile(int edges) {
+  if (tiled()) {
+    return;
+  }
+
+  save_geometry();
+
+  bool is_maximized = maximized();
+  if (is_maximized) {
+    wlr_xdg_toplevel_set_maximized(xdg_surface, false);
+  }
+
+  wlr_xdg_toplevel_set_tiled(xdg_surface, edges);
+  state = WM_WINDOW_STATE_TILED;
+}
+
+void View::maximize() {
+  if (maximized()) {
+    return;
+  }
+
+  save_geometry();
+
+  bool is_tiled = tiled();
+  if (is_tiled) {
+    wlr_xdg_toplevel_set_tiled(xdg_surface, WLR_EDGE_NONE);
+  }
+
+  wlr_output* output = wlr_output_layout_output_at(server->output_layout,
+    server->cursor->x, server->cursor->y);
+  wlr_box *output_box = wlr_output_layout_get_box(server->output_layout, output);
+
+  x = output_box->x;
+  y = output_box->y;
+
+  wlr_xdg_toplevel_set_maximized(xdg_surface, true);
+  resize(output_box->width, output_box->height);
+
+  state = WM_WINDOW_STATE_MAXIMIZED;
+}
+
+void View::window(bool restore_position) {
+  if (windowed()) {
+    return;
+  }
+
+  if (tiled()) {
+    wlr_xdg_toplevel_set_tiled(xdg_surface, WLR_EDGE_NONE);
+  }
+
+  if (maximized()) {
+    wlr_xdg_toplevel_set_maximized(xdg_surface, false);
+  }
+
+  wlr_xdg_toplevel_set_size(xdg_surface, old_width, old_height);
+
+  if (restore_position) {
+    x = old_x;
+    y = old_y;
+  } else {
+    wlr_box geometry;
+    extents(&geometry);
+
+    float surface_x = server->cursor->x - x;
+    float x_percentage = surface_x / geometry.width;
+    float desired_x = old_width * x_percentage;
+    x = server->cursor->x - desired_x;
+  }
+
+  server->damage_outputs();
+  state = WM_WINDOW_STATE_WINDOW;
+}
+
+void View::extents(wlr_box *box) const {
+  wlr_xdg_surface_get_geometry(xdg_surface, box);
+}
+
+void View::activate() {
+  wlr_xdg_toplevel_set_activated(xdg_surface, true);
+}
+
+void View::notify_keyboard_enter() {
+  wlr_keyboard *keyboard = wlr_seat_get_keyboard(server->seat);
+  wlr_seat_keyboard_notify_enter(server->seat, xdg_surface->surface,
+    keyboard->keycodes, keyboard->num_keycodes, &keyboard->modifiers);
+}
+
+void View::resize(int width, int height) {
+  wlr_xdg_toplevel_set_size(xdg_surface, width, height);
+}
+
+wlr_surface* View::surface_at(double sx, double sy, double *sub_x, double *sub_y) {
+  return wlr_xdg_surface_surface_at(xdg_surface, sx, sy, sub_x, sub_y);
+}
+
+View* View::parent() const {
+  return server->view_from_surface(xdg_surface->toplevel->parent->surface);
+}
+
+bool View::is_child() const {
+  bool is_child = xdg_surface->toplevel->parent != NULL;
+  return is_child;
 }
